changed a few hardcoded numbers to enums (better, but still ugly)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6029 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Turn Based Input.cpp
@@ -4461,7 +4461,7 @@ void GetKeyboardInput( UINT32 *puiNewEvent )
 					
 						//Heinz: 28.02.09 BUGFIX: giving G41 to Maria only in cheat mode
 						// Create object and set
-						CreateItem( (UINT16) G41, 100, &Object );
+						CreateItem( G41, 100, &Object );
 
 						pSoldier = FindSoldierByProfileID( MARIA, FALSE );
 
@@ -4951,7 +4951,7 @@ void MakeSelectedSoldierTired()
 	INT32 usMapPos;
 	if ( GetMouseMapPos( &usMapPos ) )
 	{
-		CreateItem( (UINT16)TNT, 100, &gTempObject );
+		CreateItem( TNT, 100, &gTempObject );
 		AddItemToPool( usMapPos, &gTempObject, -1, 0, 0, 0 );
 	}
